agent app - create order

+++ agent/dto/productDTO.go
@@ -3,5 +3,5 @@ package dto
 type ProductDTO struct {
 	Name            string        `json:"name"`
 	Quantity        uint          `json:"quantity"`
-	PricePerItem    float32       `json:"pricePerItem"`
+	PricePerItem    float64       `json:"pricePerItem"`
 }
