+++ agent/dto/updateProductDTO.go
@@ -4,5 +4,5 @@ type UpdateProductDTO struct {
 	ProductId		uint		  `json:"productId"`
 	Name            string        `json:"name"`
 	Quantity        uint          `json:"quantity"`
-	PricePerItem    float32       `json:"pricePerItem"`
+	PricePerItem    float64       `json:"pricePerItem"`
 }
